requirements.txt
lxml==4.8.0
packaging==21.3
pikepdf==5.1.2
Pillow==9.1.0
PyMuPDF==1.19.6
pyparsing==3.0.8
